Alpha/RPCs: change `managerPubkey` to underscores in json field names
This renames a field in the JSON format of origination operations to
make it consistent with the others.

Third party software must be updated to support this change.
For applications that just consume this format, it is enough to just recognise
both field names during the protocol change period. For applications that send
this format to the node, it is necessary to send the right one depending on the
`next_protocol` field of the interrogated block.