Alpha/RPCs: fix wrong encoding name fields
This patch fixes a mislabeled field in the JSON encoding of operation receipts.
The balance updates had a field `level` that actually contained a cycle.
It is now named `cycle`.

Third party software that make use of receipts such as block explorers
or wallets may need to be updated accordingly. A simple solution is to
support both during the protocol change period.