Alpha/Vote: new constant `test_chain_duration`
This patch turns the test chain lifetime from a hardcoded two-days
constant in the code to a parameter. This is useful for testing of
the voting system, for testnet rehearsals, or for more easily changing
the value in a future protocol upgrade.

The value of this constant is also updated to prolong the test chain
to the end of period. The test chain is now spawned at the beginning of
the third voting period and lasts the whole period.