Alpha/Michelson: do not allow spendable smart contracts
This patches forbids the creation of spendable smart contracts, a
current feature that is not necessary (it can be implemented from the
contract's code) and prevents some possible future optimisations.

It also prevents the creation of non spendable non scripted contracts,
which was mostly a design flaw allowing people to lock funds.

This is done by checking the flag in the origination operation rather
than removing it so the change is lighter for third party software.