P2p: Nack send the motive for the connection rejection.
This patch provide a `Nack` with an error code, in the vein of http
error codes that allow the client to know why its connection is not
accepted by the node.

It should help to understand the failure to connect to a node given
via RPC for example.

The P2p_rejection file is in lib_base as the rejection error will be used in
Network_version in case no network version can be negotiated.

The rejection motive is encoded as an integer. Each new motive
will have a higher number and thus fall into Unknown_argument case.

When in private mode, `raw_authenticate` of an untrusted peer result in
the P2p_errors.Private_mode error, preventing the reply with a Nack
and a list of known points.
`authenticate` function will close the connection immediately.

This patch does not change the p2p version as it is meant to be
released together with "P2p: `Nack` with a list of known peers".