Proto: remove default constants store all constants in the context
This patch does not change the semantics.

It changes the way protocol parameters (such as the block delay or the
gas limit) are provided at protocol activation and stored in the
context.

Previously they were using an optional scheme, which allowed to launch
a test network with the default mainnet constants, only specifying the
updated ones. This feature led to confusion and several mishaps on the
test networks, so it was decided that the parameters should now be
provided exhaustively in the activation block of test networks or
sandboxes.