P2p : random and parallel kill strategy on too_many_connection
Previous strategy was to kill sequentially the peers in an order
driven by the peer id. The same node A could thus get killed over and over
if the node B it connects to is overloaded (max_incoming_connections is
reached) and the node A insists on comming back.

Now the kill list is randomly selected.

Bonus: the disconnections now happen in parallel.