Proto: freeze endorsement deposits at operation application
In Athens and before, endorsement deposits where taken at the end of
the block validation, after the transactions, including transaction
from the endorsers' accounts. This made things more difficult for the
baker, and led to a few mishaps in the past.

This patch changes that behaviour, so that endorsement deposits are
taken before transactions are evaluated.