Proto: participation EMA and min/max quorum caps
Change the formula from quorum update on vote period to
participation exponential moving average. Current quorum
storage is removed and new storage participation EMA
is introduced.

Minimum and maximum quorum caps are added to the constants
of the economic protocol. Whenever a voting period would
cause the quorum to go below or above the caps it will be
bound to the limit defined in the constants.

In the future token holders can easily modify the caps by
changing the constants.