Proto: participation EMA and min/max quorum caps
Contains a BREAKING CHANGE (see end of message).

Change the formula from quorum update on vote period to participation
EMA (exponential moving average). Current quorum storage is removed
and new storage participation EMA is introduced.

Minimum and maximum quorum caps are added to the constants of the
economic protocol. Whenever a voting period would cause the quorum to
go below or above the caps it will be bound to the limit defined in
the constants.

In the future token holders can easily modify the caps by changing the
constants.

Third party developers should make sure they can parse the new fields
in the `/constants` RPC, or at least ignore them.