P2p_test: allowing to set process name prefix
This allows to set the prefix of the name of a forked process, which
serves to identify the log lines belonging to this process in addition
to its peer id.

Typical use is in the module Overcrowded where some process are
either the `Target`, a `Client_V1`or a `Client_V2`.