Proto/Migration: migrate KT1 with and without script
All spendable, scriptless contracts are migrated to 'manager.tz' script,
which replaces their functionality. It allows for the contract's manager to set
and withdraw delegate, spend the contract's funds and to set a new manager,
which is written into script's storage.

The 'manager.tz' script's parameters have field annotations, which in
combination with script entry-points allows for friendlier commands for
running the script.

Spendable and delegatable flags are replaced with entrypoints in
scripted contracts.

Co-authored-by: Benjamin Canou <benjamin@canou.fr>