Proto: add Michelson code stubs to replicate manager operations on KT1s
Spendable, scriptless contracts are simulated by the 'manager.tz' script,
which replaces their functionality. It allows for the contract's manager to set
and withdraw delegate, spend the contract's funds and to set a new manager,
which is written into script's storage.

The 'manager.tz' script's parameters have field annotations, which in
combination with script entry-points allows for friendlier commands for
running the script.

Spendable and delegatable flags are simulated by adding entrypoints to
a scripted contract.