Proto/Michelson: disable storing values of type `contract t` in newly originated contracts
Contains a BREAKING CHANGE (see end of message).

In Athens and before, Michelson contracts could store typed handles to
contracts in their storage or in constants in the code. This meant
that typechecking a contract required accessing other contracts from
the chain context. This extra type safety was not worth the
engineering cost for tooling and high level languages. Contracts will
now have to store values of type `address` and use the `CONTRACT`
instruction to typecheck contract references on demand.

All existing contracts that used the feature will continue to work
as-is. This is done by introducing a `legacy` flag througout the
typechecking code, with the following trivial semantics:
 - everything already in the chain is considered `legacy` and can
   use deprecated features,
 - everything added to the chain (parameters of transactions and code
   and storage of originations cannot.

Smart contract developers should adapt their code to store `address`es
and use instruction `CONTRACT`.