Proto/Michelson: extend comparison to linear pair structures
Michelson's `COMPARE` instruction can currently only compare simple
values (`string`s, `int`s, etc.). This limitation also applies to
`set`, `map` and `big_map` indexes.

This is an issue in particular for `big_map`s that cannot be nested,
because it prevents indexing data by a pair of indexes, such as a
`key_hash` and a `string`.

This patch lifts that restriction, allowing to compare `pair`s of
values, as long as their left component remains a simple value,
impicitly making comparable values right combs whose leaves are simple
values. The ordering is naturally lexicographic.

This patch also refactors a bit the comparison code to simplify it and
reduce code duplication.