Proto/Client: operate transfer through manageable KT1
This commit make the client capable of handling the transfers from a
KT1 with a "do" entrypoint.

It mimmic as much as possible the behaviour of the client pre-Babylon
on spendable KT1. However, fees are taken from the contract manager,
and the ̀source`of the operation will be the manager, not the KT1.