Doc/Michelson: Fix the documentation of the (UN)PAPAIR macros
The previously documented unfolding matched neither the implementation
nor the expected semantics for these macros.

The new unfolding has the correct semantics but it still does not
match the implementation, which is more complex to explain.

It is not clear to me whether the current implementation brings any
benefit with respect to the simple unfolding that is documented here
so I do not know if we should change the implementation or the
documentation.