Protocol/Michelson: fix the interpretation of the MAP instruction on maps
In the previous implementation, accumulating a value during a MAP on a
map was impossible because the initial stack tail was restored. This
was not the documented behavior of the MAP instruction and it was
inconsistent with the case of mapping over a list.

BREAKING CHANGE: originated contracts that rely on the previous and
incorrect) semantics might behave incorrectly.