Proto/Env: handle existing value in in-memory context copy function
As it is, when `Memory_context.copy` returns `None`, `Raw_context.copy` will
fail with `Missing_key` error. However, when the value is already set,
`Memory_context.copy` also returns `None`, which makes this fail with a
misleading error message.

Although in `src/lib_storage/context.ml` copy implementation the value may
already exists, it seems useful to make it fail in in-memory context.

To fix this, we make the in-memory context copy function fail when the value
is already set.