Proto: Allow to apply balance updates from migration
This commit allows to attach balance update receipts from a migration.
Because the migration runs before the first block of the new protocol,
we cannot return balance updates directly from migration. Instead, we add the
balance updates receipts when we're finalizing the first block.

This adds a new storage (`Storage.Pending_migration_balance_updates`) that can
be set during migration and checked when we try to finalize a block to apply
balance updates receipts.