Proto/Client: no partial entrypoints list when typecheck fails
To detect the presence of a "default" entrypoint and get its type, the
current implementation uses the entrypoint typechecking RPC. If it
fails, it assumes that the entrypoint doesn't exists.

This patch discriminate the 404 Not Found error that indicates the
absence of such entrypoint with other error that can happen, for
example Forbidden access in presence of RPC filters, or internal error
in case of proxy malfunction.