Tests/python: check node connectivity (improvements)
When running a node with `Node.run`, it may take some time before
the node is accepting RPC connections.

Currenty, an arbitrary sleeping time is enforced in `sandbox.add_node`
and in the tests that restart nodes. This may fail non-deterministically
on some setups.

This patch and the one before that add a new method to the client that
polls a node using some RPC. `time.sleep` are replaced by a call to
this method.

This should make tests more robust and less dependent on the execution
platform.