Registration: move module to fix type dependency
Registration refers to `Binary_schema`. It used to refer to
`Data_encoding__Binary_schema.t` which is the internal reprensentation.
It now refers to `Data_encoding.Binary_schema.t` which is the
reprensentation intended for users of the library.

A very basic test also checks that the registration system behaves as
expected.