Flextesa: increase the number of attemps to start the node
Before, the client took ~1s to start.
With 8 attempts, it thus added 8s to the delay we allowed the node to start.
Now that the client starts almost instantly, we gave 8 less seconds to the node,
even though it only starts 1s faster.

By setting the number of attempts to 12, we add about 20s allowed delay.