p2p: Re-implement PeerGreyList with a LRU cache
Replaces the implementation of the PeerTable module based on the Ring
Buffer, with one based on LRU caches. The changes provide LRU
semantics to the Peer Greylist, and avoid the inconsistencies that
could arrive using the Ring Buffer, e.g. stuffing the ring with copies
of Peer.IDs which have already been removed from the greylist.