p2p,bin_node: add the concept of expected peer identity
An optional peer identity can be specified with a point. In particular
the `--peer` option of the node now can take an expected peer id using
the b58 format: './tezos-node rune --peer <address>#id'. If an id is
specified, the node checks that this id is the same as the one sent by
the peer when we connect to it.

In case where the peer identity is specified while we are already
connected to a node (see the next commit), if the expected identity
is different from the current identity, we disconnect from this point.