Tests michelson, Doc: more idiomatic test of opcodes
CAR/CDR : add subtype to params to improve documentation
Adding different subtypes to contracts parameters allows to follow the
semantics of opcodes CAR and CDR just by looking at the typing trace.
The type of the stack after the operation now carry the information
about which element (left or rght) has been taken.

CHECK_SIGNATURE: more idiomatic use
The operation already check the signature against the hash of
the bytes passed, so no need to hash them beforehand.
The semantics of the program is not quite the same but it is a better
example of how to use the opcode.

CREATE_CONTRACT: more idiomatic
The example was not issuing the operation created by CREATE_CONTRACT,
so no contract was originated. It is confusing for the opcode
documentation, where we rather want to explain how to originate a contract.