Clic: Add non-terminal sequence of parameters
This adds a sequence that doesn't have to be the last parameter. It is followed
by suffix keywords and further parameters. On execution, the suffix will be
attempted to be matched before the sequence's parameters. Because of this,
care must be taken that the sequence's parameters cannot be interpreted as the
suffix keywords. For example, it shouldn't be used with parameters that can be
arbitrary strings such as aliases.