Shell/Workers: limit logger instantiation
Specifically, this MR instantiates the Event_definition in
worker_loggers per Worker functor instantiation (rather than once per
worker). This fixes a small memory leak.

In order to accomplish the fix described above, worker_logger exports
strings rather than a type that is abstract outside of the instantiated
module. The strings are lazy in case the logging is disabled and
formatting does nothing.

Co-authored-by: Raphaël Proust <code@bnwr.net>
Co-authored-by: MBourgoin <mathias.bourgoin@tezcore.com>