Client/Michelson: add minimal client support for TZT
This commit adds a `run unit tests` command in the client which can be
used to run a set of unit tests similarly to how `typecheck script`
can be used to type check that a set of scripts.

The support for the TZT format added in this commit is very minimal;
only the mandatory primitives `input`, `output`, and `code` are
recognized.