ProtoEnv/V1: disallow concurrent list processing
This affects both Error_monad and Lwt/Lwt_list and removes some
functions that encourage concurrent processing of lists. The aim is to
reduce non-determinism.

Technically, using concurrent processing of lists if the functions
applied to the list elements are pure and cannot fail is ok. But there
is no static way to guarantee that it is the case.