block_locator: add some comments and remove the raw type
The raw type seems useless because the type t is private. Moreover,
this type was used only once in the `peer_validator` module. The
explicit coercion as been replaced by a coercion at the type level
as it was done several lines above.

Some comments in the head of a the `block_locator` module are added
to explain how it was used.