Client/mockup: create base_dir in mockup mode when it is acceptable
There were two layers of identical checks: one in
create_mockup_command_handler and another one in create_mockup itself. I
combined the checks so that they are now only in create_mockup.

The function classify_base_dir contains all the logic for analysis of the
state of base_dir, so I re-used it in create_mockup. Now we do create
base_dir if it is missing but only when it did not exist or was empty. These
conditions should be enough to prevent it from overwriting existing data. In
addition to that we indicate the fact that a new directory was created by
printing a message.

I also made classify_base_dir more precise: it now recognizes another
pathological case when base_dir is a file. It looks like the old
implementation was buggy: when given a file it would declare that the
directory does not exist.

I also renamed mockup_mode into client_mode, because it includes
Mode_client, which has nothing to do with mockup.