CI: compute/publish code coverage
Adds two jobs to the CI, which can be launched manually.

The test_coverage job launches the test suite (sequentially) and
generates an html report. The report artefact can be downloaded
from Gitlab CI web interface.

Alternatively, the pages job publishes this report on the static website
linked to the repo. It is accessible via a URL of the form

$CI_PAGES_URL/$CI_PIPELINE_ID

Examples:
    https://nomadic-labs.gitlab.io/tezos/105822404/
    https://tezos.gitlab.io/tezos/1234822404/