internal_event: Fix the mutable variable `section`.
A reference called `sections` was used to register the different
sections of the code so that they can be printed afterwards in the
manual of `tezos-node run`. However, this reference was not updated
with the new events modules such as `Simple` or the one from the
`Worker_logger`.

We have also renamed this variable to `registered_sections`.