Shell/bootstrap: spawned chain inherits bootstrap flag from parent
The bootstrap heuristic doesn't deal well with the corner
case where all nodes spawn a chain at a same time, especially
if the (global) bootstrap threshold isn't zero for the node which
bake the first blocks.

This solves the issue by making the testchain inherit the bootstrap
status from the parent.