Client/mockup: Use uri_encoding from RPC_encoding, not Client_keys
This is a minor* fix.

`uri_encoding` was already available from the module interface of
`RPC_encoding`.

There is thus no need to export it from `Client_key`.

As an added benefit, we might, without explicitly asking for it, have preserved
a designed abtraction in `Client_keys`. You never know ;-)