Client/mockup: create mockup directory after successful parsing
This reorders the phase of the mockup creation so that directory creation occurs
at the end of the whole setup. The goal is to fail early if the base directory
is at odds with the requirements of mockup mode.

Previously, the mockup directory was created before parsing, always.
 So that launching

./tezos-client --base-dir /tmp/mockup -M mockup create mockup --protocol-constants /tmp/bad-sandbox-parameters.json

with syntactically bad parameters would create the directory, but not the
correct files, thus resulting in the following error during a future call:

./tezos-client --base-dir /tmp/mockup -M mockup create mockup --protocol-constants /tmp/good-sandbox-parameters.json
Error:
  /tmp/mockup-base is not empty, please specify a fresh base directory

with syntactically correct parameters.

The same would happen with `--bootstrap-accounts` option.