Test: extract client instanciation in sandbox
For the moment, only one client is used in the different tests, through sandbox.client(0). For a given node, only this client is used, including its state (saved keys, known addresses, etc). It may be interesting to allow instantiating a new client for a given node for scenari requiring clients with different states.