Lib_p2p/tests: tracking detached function results
Detached process results were not available from the main process,
this patch gives access to the detached functions results when the
process is not killed or canceled.

It also bring the possibility to use data-encoding for exchanges with
the detached process.

As it is not yet possible to define a parallel error that would relies
on the registration of recursive errors, the head of each error trace
is taken in the trace when joining multiple process.