Client/mockup: Add chain-id-seed cli parameters
The chain id seed on the command line allows the user to specify a value that
will generate a valid chain identifier for the mockup experiments.

The generated chain id is printed back to the user, e.g., to be used later
inside a JSON file specifying a bunch of protocol parameters, including the
chain id.