Makefile: only list protocols with a TEZOS_PROTOCOL file
If a protocol is deleted or moved, its `src/proto_XXX` directory
will probably remain when switching branch because of leftover files
which are ignored by Git. This is particularly annoying for users
which just want to `git pull && make`, as they face an error
`No rule to make target 'src/proto_plop/lib_protocol/TEZOS_PROTOCOL`.

This patch modifies the Makefile to only try to build protocols for which
there is actually a `TEZOS_PROTOCOL` file, instead of all `proto_*` folders.