Tests/python: --sandbox is no longer needed
The `--sandbox` option is no longer needed when launching a node.
It was used to specify the public key to be used by genesis to
validate a protocol activation.

This key is now hard-coded in the node and is used by default when
the node is launched with `--network=sandbox`.

This commit updates the python test API and tests to reflect this
change.

There's also no need for `--network` in snapshot import