Tests/P2P: avoid use of Alcotest_lwt
... in favour of explicitly calling `Lwt_main.run` for each threaded
test case – replicating the semantics of pre-1.0 Alcotest_lwt.

These tests nest calls to `Lwt_main.run` as part of forking nodes, which
is undefined behaviour pre-5.0.0 and explicitly forbidden thereafter.
When the outer top-level promise contains the full test-suite, as is the
case for Alcotest_lwt 1.0+, this results in each fork triggering its own
suite of tests (which then deadlock due to contention over sockets).