lib_shell: Factorize the timestamp drifting check
For synchrnoisation issues, the shell accepts blocks which are not too
far away in the future (about 15s). We propose in this MR to factorize
the use of this constant in a new module `Clock_drift.ml`. We also
factorize the function `soon` which is actually used three times.

The use of this constant can be checked with `rgrep "15" *.ml` on the
files of the `Tezos` directory and especially the ones of `lib_shell`.