Proto: Add storage for baker consensus key
Baker consensus key is a public key that can be used for baking and endorsing
blocks.

It's extended for compatibility with the current protocol version by
establishing 1-to-1 mapping between baker contracts and consensus keys and
enforcing that the implicit account of a consensus key must be unallocated
(that is have no balance). Instead, transfers to and from a consensus key
are mapped to the baker account that uses the given key.

This allows the consensus key to:
- sign operations using the consensus key
  - `Transfer` operation from the baker contract
  - `Proposals` and `Ballot` operation
- allow the consensus key to continue to work as a target of
  - `Delegation` operation and `SET_DELEGATE` instruction
  - `Origination` operation and `CREATE_CONTRACT` instruction

Each baker contract has an active consensus key and possibly a pending
consensus key.

An active baker consensus key is stored in snapshot-able storage
(`Storage.Baker.Consensus_key`) together with a reverse map
(`Storage.Baker.Consensus_key_rev`). An active consensus key value is read from
a snapshot for the current cycle. The snapshots are cleared after
`Constants.preserved_cycles` to allow the protocol to check historical values
for double baking and double endorsement evidence.

A pending consensus key is a paired with its activation cycle. It can be
updated at any point. This will set the key as pending for the given baker
(`Storage.Baker.Pending_consensus_key`) and set its activation cycle
`Constants.preserved_cycles + 2` in future.

A pending consensus key will become active in the beginning of the activation
cycle.

If a new consensus key is set before the current one is activated, the current
pending key will be overridden and its activation cycle is reset.