CI: Remove `test_voting_power_opcode.py` from integration tests
Alvaro said:
If I remember correctly, this test vanished after we decided to perform it within the regression tests. The original test was dynamic in its inputs were not fixed, but retrieved from the account's balances. These inputs had to be fixed in the regression tests in order to accommodate to the testing framework.