Tests/python: simplify mockup client tests
This commit does the following:

1/ It fixes disk leaks in the mockup's tests. Indeed
   each successful execution of the mockup's tests was
   leaving **9** dangling directories named /tmp/tezos-client*
2/ It gets rid of client.py's set_base_dir method
   which relies on a side effect and hence makes
   mockup initialization harder to understand.
3/ Simplifies initialization in tests of the mockup
   It was cumbersome for bad reasons before, mostly
   because it was authored by two tezos noobs (one of them
   being my earlier myself)

This commit removes the test
that `tezos-client --base-dir /tmp/file create mockup`
fails when /tmp/file is a file. This cannot be done anymore because
client.py's Client class defensively checks that the base directory is
a directory so an exception is raised even before calling `create mockup`
It's okay, it's not a super important test.