Lib_p2p/test: remove nested Lwt_main.run in detached process
Nesting of Lwt_main.run is forbidden since Lwt 5.
We used this nesting to prevent the yielded promises to be rescheduled
in the detached process, which results a kind of fork-bomb.

As it is not possible anymore, we have to run the forks sequentially
to ensure that there is no yielded promises waiting to fork inside the
forked process.