Shell: Store and validate operation metadata hashes
We store operation metadata hashes for blocks starting from environment V1.
These are returned from the block validation.

The context commits to the root hash of predecessor's operation metadata to
prepare to change blocks committing to the context as of the end of previous
block, which will allow to bake blocks without having to execute transactions.