P2p: fix incoming count
There was a window between the removal of a point from the incoming
table and the time where the file descriptor was closed or accounted
in active connection.
Accumulation of incoming connection in this window would lead to
exhaustion of available file descriptor and to node's death.

We also shouldn't crash the node because an exception arised in the
handling process of a specific connection.