Tests/python: mockup: tests both alpha and carthage
This is done by parameterizing the mockup_client fixture by the
two protocols, as well as parameterizing the tests that do
not use the fixture.

On the way, I moved the mockup_client fixture to conftest.py
as it doesn't hardcode the protocol anymore (this was the blocking point
for making it widely available).