Tests/python: make carthage tests pass
We specify `--protocol` because it is required for `config show mockup`
and `config init mockup` to work. Previously it was working
by luck, because when `--protocol` is not specified, the client
will take the first proxy implementation, which happened to be alpha.
So it was working until we started testing carthage.