Emacs/michelson-mode: update header to comply with Emacs convention
Emacs Lisp files that are meant to be distributed are expected to have
a header following the convention described in the Emacs Lisp
manual (https://www.gnu.org/software/emacs/manual/html_node/elisp/Simple-Packages.html#Simple-Packages).

The main advantage of following this convention is that the header
fields are understood by the Emacs package manager so that
michelson-mode can be installed with the `package-install-file`
command.