Client/mockup: rewrite RPC_client
The point of this rewrite is that `generic_json_call_stub` and
`call_service_local` did not rely on common smaller functions (like resto's
http client does with `prepare` and `generic_call`) . As a consequence,
`call_service_local` was failing to find the
/chains/chain_id/mempool/pending_operations RPC (we had a temporary
workaround that made `call_service_local` delegate to
`generic_json_call_stub`, this rewrite is the lasting solution).

This commit rewrites the mockup's RPC client so that it follows
the structure of resto's http client.

This also corrects the implementation of
call_streamed_service_local which was too simplistic. For example the call to
on_chunk should be conditional (see the implementation of
`call_streamed_service` in resto-cohttp-client's client.ml file)