Client/mockup: baking add trashpool for refused transactions
In mockup mode, a refused transaction is forever unbakable since the mockup
chain has a single live block, the head, which has just moved with a newly baked
block. That means, in particular, that the branch the refused transaction is
referring is not live anymore, so that there's no chance for this transaction to
be addded later on. In effect, the mockup "chain" has a ttl of 0 for only the
present block is okayed. To help the user, the refused transactions are set
aside in a so-called trashpool for review.