Scripts: adapt pre-commit hook to poetry
This adapts the pre-commit hook so that it uses correctly the following
entries from `tests_python/Makefile`:

```
TYPECHECK=poetry run mypy
LINT=poetry run pylint
LINT2=poetry run pycodestyle
```

Before the right-hand side of these lines was a single word (such as
`mypy`) and the pre-commit hook was hence considering the whole
right-hand side as a command.

With this commit, the pre-commit hook considers the right-hand sides
as space-separated commands, hereby tokenizing "poetry run mypy"
to ["poetry", "run", "mypy"] and using that as the command to execute
programmatically.

This commit also takes care of executing `pytest` from the
`tests_python` directory, as this is a requirement for `poetry run` to
work.

This commit was part of
https://gitlab.com/tezos/tezos/-/merge_requests/1711 initially
but was ultimately inadvertently left out. Hence, here it is.