Tests/python: add test generating OpenAPI specifications
The test mimicks what src/openapi/generate.sh does.

Because the test reuses sandbox.py facilities, it doesn't have to
start a node nor a client on its own; as opposed to generate.sh.

Ideally, when the python tests framework becomes a library, generate.sh
should be deleted or modified; so that it reuses the test's core logic.